+++ MordoEngine/src/Scene/Game/GameScene.cpp
@@ -2,11 +2,15 @@
 #include "../../Core/Managers/Manager.h"
 #include "../../Core/Shader/Shader.h"
 #include "../../Input/Input.h"
+#include "../../Renderer/SkyBoxRenderer.h"
 
-GameScene::GameScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Camera> camera, std::shared_ptr<Renderer> renderer) 
-	: Scene("terrain"), m_Terrain(terrain) , m_Camera(camera) , m_Renderer(renderer) , m_Sun(std::make_unique<Sun>(0.9f))
+GameScene::GameScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Camera> camera, std::shared_ptr<Renderer> renderer)
+	: Scene("terrain"), m_Terrain(terrain), m_Camera(camera), m_Renderer(renderer), m_Sun(std::make_unique<Sun>(0.9f))
 {
 	m_CameraController = std::make_unique<GameCameraController>(m_Camera);
+	m_SkyBoxRenderer = std::make_unique<SkyBoxRenderer>(
+		Manager<Shader>::Get("skyBox")
+	);
 }
 
 void GameScene::OnEntry()
@@ -23,20 +27,29 @@ void GameScene::Update(float deltaTime)
 
 void GameScene::Render()
 {
-	glm::mat4 projection = m_Camera->GetProjectionMatrix();
-	glm::mat4 view = m_Camera->GetViewMatrix();
-	glm::mat4 model = glm::mat4(1.0f);
-	glm::vec3 lightDir = m_Sun->GetReverseLightDirection();
-	m_Renderer->Render(
-		m_Camera->GetPosition(),
-		&view,
-		&projection,
-		&model,
-		&lightDir
-	);
+    glm::mat4 projection = m_Camera->GetProjectionMatrix();
+    glm::mat4 view = m_Camera->GetViewMatrix();
+    glm::mat4 model = glm::mat4(1.0f);
+    glm::vec3 lightDir = m_Sun->GetReverseLightDirection();
+
+    m_Renderer->Render(
+        m_Camera->GetPosition(),
+        &view,
+        &projection,
+        &model,
+        &lightDir
+    );
+
+    m_SkyBoxRenderer->Render(
+        &view,
+        &projection,
+        nullptr,
+        nullptr
+    );
+
 }
 
 GameScene::~GameScene()
 {
-	
+
 }